requirements.txt
asgiref==3.8.1
autopep8==2.3.2
dj-database-url==2.2.0
Django==5.2.1
django-cors-headers==4.6.0
django-environ==0.12.0
django-filter==25.1
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
flake8==7.2.0
gunicorn==22.0.0
isort==5.13.2
psycopg2==2.9.10
pycodestyle==2.13.0
pyflakes==3.3.2
pylint==2.17.7
pylint-django==2.5.3
redis==6.1.0
sqlparse==0.5.3
whitenoise==6.8.2
